Refactored time series module
