+++ frontend/src/app/modules/time-series/components/time-series-chart/time-series-chart.component.scss
