+++ frontend/src/app/modules/time-series/models/context-menu-position.model.ts
@@ -1,8 +1,8 @@
-import {TimeSeriesPoint} from "./time-series-point.model";
+import {TimeSeriesPoint} from './time-series-point.model';
 
 export default class ContextMenuPosition {
-  //if divider is set on true, the other fields are ignored
-  divider?: boolean = false;
+  // if divider is set on true, the other fields are ignored
+  divider ? = false;
 
   title?: string;
   icon?: string;
