+++ frontend/src/app/modules/time-series/models/event-result-data.model.ts
@@ -2,14 +2,14 @@ import {EventResultSeriesDTO} from './event-result-series.model';
 import {SummaryLabel} from './summary-label.model';
 
 export interface EventResultDataDTO {
-  measurandGroups: {[key: string]: string};
+  measurandGroups: { [key: string]: string };
   series: EventResultSeriesDTO[];
   summaryLabels: SummaryLabel[];
   numberOfTimeSeries: number;
 }
 
 export class EventResultData implements EventResultDataDTO {
-  measurandGroups: {[key: string]: string};
+  measurandGroups: { [key: string]: string };
   series: EventResultSeriesDTO[];
   summaryLabels: SummaryLabel[];
   numberOfTimeSeries: number;
