+++ frontend/src/app/modules/time-series/models/event-result-point.model.ts
@@ -1,4 +1,4 @@
-import {WptInfo} from "./wpt-info.model";
+import {WptInfo} from './wpt-info.model';
 
 export interface EventResultPointDTO {
   date: Date;
