+++ frontend/src/app/modules/time-series/services/chart-feature-services/line-chart-draw.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { LineChartDrawService } from './line-chart-draw.service';
+
+describe('LineChartDrawService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: LineChartDrawService = TestBed.get(LineChartDrawService);
+    expect(service).toBeTruthy();
+  });
+});
