+++ frontend/src/app/modules/time-series/services/chart-feature-services/line-chart-scale.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { LineChartScaleService } from './line-chart-scale.service';
+
+describe('LineChartScaleService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: LineChartScaleService = TestBed.get(LineChartScaleService);
+    expect(service).toBeTruthy();
+  });
+});
