+++ frontend/src/app/modules/time-series/time-series.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit, ViewEncapsulation} from '@angular/core';
-import {URL} from "../../enums/url.enum";
-import {LinechartDataService} from "./services/linechart-data.service";
-import {ResultSelectionStore} from "../result-selection/services/result-selection.store";
+import {URL} from '../../enums/url.enum';
+import {LineChartDataService} from './services/line-chart-data.service';
+import {ResultSelectionStore} from '../result-selection/services/result-selection.store';
 import {EventResultData, EventResultDataDTO} from './models/event-result-data.model';
 import {BehaviorSubject} from 'rxjs';
 
@@ -10,7 +10,7 @@ import {BehaviorSubject} from 'rxjs';
   templateUrl: './time-series.component.html',
   styleUrls: ['./time-series.component.scss'],
 
-  //used to render context menu with styles from time-series.component.scss file
+  // used to render context menu with styles from time-series.component.scss file
   encapsulation: ViewEncapsulation.None
 })
 export class TimeSeriesComponent implements OnInit {
@@ -18,7 +18,8 @@ export class TimeSeriesComponent implements OnInit {
   public showTimeSeriesChart = false;
   public results$ = new BehaviorSubject<EventResultDataDTO>(new EventResultData());
 
-  constructor(private linechartDataService: LinechartDataService, private resultSelectionStore: ResultSelectionStore) { }
+  constructor(private linechartDataService: LineChartDataService, private resultSelectionStore: ResultSelectionStore) {
+  }
 
   ngOnInit() {
     if (this.resultSelectionStore.validQuery) {
