+++ frontend/src/app/modules/time-series/time-series.module.ts
@@ -4,9 +4,9 @@ import {TimeSeriesComponent} from './time-series.component';
 import {RouterModule, Routes} from '@angular/router';
 import {SharedModule} from '../shared/shared.module';
 import {ResultSelectionModule} from '../result-selection/result-selection.module';
-import {TimeSeriesLineChartComponent} from './components/time-series-line-chart/time-series-line-chart.component';
+import {TimeSeriesChartComponent} from './components/time-series-chart/time-series-chart.component';
 import {LineChartService} from './services/line-chart.service';
-import {LinechartDataService} from './services/linechart-data.service';
+import {LineChartDataService} from './services/line-chart-data.service';
 import {FormsModule} from '@angular/forms';
 
 const TimeSeriesRoutes: Routes = [
@@ -16,7 +16,7 @@ const TimeSeriesRoutes: Routes = [
 @NgModule({
   declarations: [
     TimeSeriesComponent,
-    TimeSeriesLineChartComponent
+    TimeSeriesChartComponent
   ],
   imports: [
     CommonModule,
@@ -27,7 +27,7 @@ const TimeSeriesRoutes: Routes = [
   ],
   providers: [
     LineChartService,
-    LinechartDataService
+    LineChartDataService
   ],
   exports: [
     TimeSeriesComponent
